requirements.txt
flask>=3.0.0
flask-cors>=4.0.0
groq>=1.0.0
python-dotenv>=1.0.0
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
